+++ examples/serverless/activity_names.py
@@ -1,0 +1,4 @@
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT License.
+
+REMOTE_HELLO = "remote_hello"